requirements.txt
datajoint
matplotlib
numpy
tensorflow
opencv-python>=4.5.1
git+https://github.com/philferriere/cocoapi.git#egg=pycocotools&subdirectory=PythonAPI
pyrender>=0.1.45
mmcv-full
mmdet
mmtrack
mmpose
python-dotenv
GitPython

# Non-packages. Required, but not installable. Section read by env.py code
# git+https://github.com/nkolot/ProHMR
# git+https://github.com/CMU-Perceptual-Computing-Lab/openpose
# git+https://github.com/Jack000/Expose
# git+https://github.com/Arthur151/ROMP
# git+https://github.com/fabro66/GAST-Net-3DPoseEstimation
# git+https://github.com/zczcwh/PoseFormer
# git+https://github.com/mkocabas/VIBE
# git+https://github.com/ZhengyiLuo/MEVA
# git+https://github.com/mkocabas/PARE
# git+https://github.com/pixie-io/pixie
# git+https://github.com/davrempe/humor
# git+https://github.com/ifzhang/FairMOT
# git+https://github.com/CharlesShang/DCNv2
# git+https://github.com/PeizeSun/TransTrack
# git+https://github.com/yaodongyu/TRADES
# git+https://github.com/paTRICK-swk/Pose3D-RIE
# git+https://github.com/facebookresearch/VideoPose3D
# git+https://github.com/jfzhang95/PoseAug